requirements.txt
amqp==5.2.0
asgiref==3.8.1
billiard==4.2.0
black==25.1.0
celery==5.4.0
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
cron-descriptor==1.4.3
Django==5.0.6
django-celery-beat==2.6.0
django-timezone-field==7.0
django_celery_results==2.6.0
flower==2.0.1
gunicorn==23.0.0
humanize==4.12.3
idna==3.7
isort==6.0.1
kombu==5.4.2
mypy_extensions==1.1.0
packaging==24.2
pathspec==0.12.1
pillow==10.3.0
platformdirs==4.3.8
prometheus_client==0.22.1
prompt_toolkit==3.0.47
psycopg2-binary==2.9.10
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2025.2
redis==5.2.1
requests==2.32.3
six==1.16.0
soupsieve==2.6
sqlparse==0.5.0
tornado==6.5.2
tzdata==2024.1
urllib3==2.2.2
vine==5.1.0
wcwidth==0.2.13
whitenoise==6.8.2
